updated jest config.. to run tests in each separate package
