+++ jest.base.config.js
@@ -3,6 +3,7 @@ module.exports = {
   coveragePathIgnorePatterns: ['(tests/.*.mock).(jsx?|tsx?)$'],
   modulePathIgnorePatterns: ['dist', 'examples', 'node_modules'],
   collectCoverageFrom: ['**/*.ts', '!**/node_modules/**', '!**/dist/**'],
+  coverageDirectory: '<rootDir>/coverage/',
   testMatch: ['<rootDir>/packages/**/tests/**/*.test.ts'],
   transform: {
     '^.+\\.ts?$': 'ts-jest',
