+++ packages/api/jest.config.js
@@ -1,7 +1,11 @@
-const baseConfig = require('../../jest.config.base');
+const baseConfig = require('../../jest.base.config');
+const packageJson = require('./package.json');
+const packageName = packageJson.name.split('@agile-ts/').pop();
 
 module.exports = {
   ...baseConfig,
   rootDir: '../..',
-  name: 'API',
+  roots: [`<rootDir>/packages/${packageName}`],
+  name: packageName,
+  displayName: packageName,
 };
